requirements.txt
# QAGen Pipeline 依赖

# 核心依赖
pydantic>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.28.0
aiohttp>=3.8.0

# Tokenization
tiktoken>=0.5.0

# OCR (PaddleOCR)
paddleocr>=2.7.0
paddlex>=3.0.0
pandas>=2.0.0
tabulate>=0.9.0  # pandas to_markdown



# LightRAG
lightrag-hku>=1.0.0

# Markdown处理
markdown>=3.4.0

# 可选: 图片OCR (Tesseract)
# pytesseract>=0.3.10

# JSONL处理
jsonlines>=3.1.0

